finalise subpackages
