+++ src/main/java/com/github/karllevik/qmorph/geom/Triangle.java
@@ -1,6 +1,6 @@
 package com.github.karllevik.qmorph.geom;
 
-import com.github.karllevik.qmorph.Msg;
+import com.github.karllevik.qmorph.viewer.Msg;
 
 /**
  * A class holding information for triangles, and with methods for the handling
