+++ src/main/java/com/github/karllevik/qmorph/viewer/MyFilterOutputStream.java
@@ -1,4 +1,4 @@
-package com.github.karllevik.qmorph;
+package com.github.karllevik.qmorph.viewer;
 
 import java.io.IOException;
 import java.io.OutputStream;
